+++ blocklylib-core/src/main/java/com/google/blockly/android/ui/WorkspaceHelper.java
@@ -195,7 +195,7 @@ public class WorkspaceHelper {
      *
      * @return The value in virtual view units.
      */
-    public int workspaceToVirtualViewUnits(int workspaceValue) {
+    public int workspaceToVirtualViewUnits(float workspaceValue) {
         return (int) (mDensity * workspaceValue);
     }
 
@@ -441,7 +441,7 @@ public class WorkspaceHelper {
      */
     public void workspaceToVirtualViewCoordinates(WorkspacePoint workspacePosition,
                                                   ViewPoint viewPosition) {
-        int workspaceX = workspacePosition.x;
+        float workspaceX = workspacePosition.x;
         if (mRtl) {
             workspaceX *= -1;
         }
@@ -526,13 +526,13 @@ public class WorkspaceHelper {
     public Rect getViewableWorkspaceBounds(Rect outRect) {
         mTempViewPoint.set(0, 0);
         virtualViewToWorkspaceCoordinates(mTempViewPoint, mTempWorkspacePoint);
-        outRect.left = mTempWorkspacePoint.x;
-        outRect.top = mTempWorkspacePoint.y;
+        outRect.left = (int) mTempWorkspacePoint.x;
+        outRect.top = (int) mTempWorkspacePoint.y;
 
         mTempViewPoint.set(mWorkspaceView.getWidth(), mWorkspaceView.getHeight());
         virtualViewToWorkspaceCoordinates(mTempViewPoint, mTempWorkspacePoint);
-        outRect.right = mTempWorkspacePoint.x;
-        outRect.bottom = mTempWorkspacePoint.y;
+        outRect.right = (int) mTempWorkspacePoint.x;
+        outRect.bottom = (int) mTempWorkspacePoint.y;
         return outRect;
     }
